Update budget backend creation

+++ server/BudgetBoard-Database/Models/Budget.cs
@@ -4,7 +4,7 @@ public class Budget
     public Guid ID { get; set; }
     public required DateTime Date { get; set; }
     public required string Category { get; set; }
-    public float Limit { get; set; } = 0.0f;
+    public decimal Limit { get; set; } = 0.0M;
     public required Guid UserID { get; set; }
     public ApplicationUser? User { get; set; } = null!;
 }
